LayoutTests/fast/loader/stateobjects/replacestate-in-iframe-expected.txt

main frame - has 1 onunload handler(s)
frame "<!--frame1-->" - has 1 onunload handler(s)
ALERT: Navigating back...
main frame - has 1 onunload handler(s)
frame "<!--frame1-->" - has 1 onunload handler(s)
ALERT: onpopstate
PASS
